# feat(query): support row fetch for merge into