I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary

1. Support row fetch for merge into.
2. Refactor merge into physical plan, add `MergeIntoSplit` and `MergeIntoManipulate` to better build pipeline.
3. Fix the bug when dividing tasks in `ParquetRowsFetcher`.

We use the `enable_merge_into_row_fetch` setting to control whether row fetch is enabled, and we will adjust it adaptively later.

## Performance Test (Databend Cloud Small)
### 1. TPC-H SF100
**Query Duration: 22.6s → 7.1s**
**Byte Scanned: 107 GB → 8 GB**
```sql
create table small_lineitem as select * from lineitem limit 500000;
create table lineitem_source as select * from (select *, ROW_NUMBER() OVER (PARTITION by l_orderkey) as rn from small_lineitem) t where t.rn = 1;
-- 124989 rows matched
select count(*) from lineitem_source as source, lineitem as target where source.l_orderkey = target.l_orderkey and source.l_linenumber = target.l_linenumber; 

-- merge into test sql
merge into lineitem as target using lineitem_source as source on target.l_orderkey = source.l_orderkey and source.l_linenumber = target.l_linenumber when matched then update set target.l_quantity = source.l_quantity + 1 when not matched then insert *;
-- main: 22.6s
set global enable_merge_into_row_fetch = 0;
merge into lineitem as target using lineitem_source as source on target.l_orderkey = source.l_orderkey and source.l_linenumber = target.l_linenumber when matched then update set target.l_quantity = source.l_quantity + 1 when not matched then insert *;

-- enable merge into row fetch: 7.1s
set global enable_merge_into_row_fetch = 1;
merge into lineitem as target using lineitem_source as source on target.l_orderkey = source.l_orderkey and source.l_linenumber = target.l_linenumber when matched then update set target.l_quantity = source.l_quantity + 1 when not matched then insert *;

```
| Main | PR |
|:----:|:--:|
| <img src="https://github.com/datafuselabs/databend/assets/54522439/499d9c31-506f-4bf9-891e-af0bfbfafd63" alt="main" width="326"> | <img src="https://github.com/datafuselabs/databend/assets/54522439/1ded0b4c-c1fa-44f4-b96e-a28282c40be9" alt="pr" width="333"> |

### 2. Table with 50 columns
**Byte Scanned:  186 GB → 7 GB**
```
target: numbers(50000000)
source: numbers(100000)
merge into target using source on target.col_0 = source.col_0  when matched then update set target.col_1 = source.col_1 + 1 when not matched then insert *;
```


<!--
Briefly describe what this PR aims to solve. Include background context that will help reviewers understand the purpose of the PR.

- Fixes #[Link the issue here]
-->

## Tests

- [ ] Unit Test
- [x] Logic Test
- [ ] Benchmark Test
- [ ] No Test - _Explain why_

## Type of change

- [ ] Bug Fix (non-breaking change which fixes an issue)
- [x] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [ ] Refactoring
- [x] Performance Improvement
- [ ] Other (please describe):

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/datafuselabs/databend/15859)
<!-- Reviewable:end -->